refactor: extract clearFeedLocationHash into a hook
